# Added `@cache` to `detype_pattern()`.