@tfranzel 2022-03-19
Hi @ngnpope,

thank you for looking into this. excellent find! I also had a look just now and it appears that there are 2 issues. 

1. `detype_pattern` is called repeatedly on all patterns for each endpoint. Basically this being O(n*m) where it could be O(m). doh!
2.  In itself `detype_pattern` is not doing anything expensive, but apparently rebuilding `URLResolver` appears to be quite expensive in the background.

Your intuition is right about caching this, but I would be more at ease if it were slightly different. Caching it once (since it is not supposed to change), while not wasting memory on cached values that are never retrieved again. This should be identical in terms of speed assuming there are no reused subtrees. Could you try and benchmark this instead?

```diff
diff --git a/drf_spectacular/plumbing.py b/drf_spectacular/plumbing.py
index 81e35f3..bbaf07d 100644
--- a/drf_spectacular/plumbing.py
+++ b/drf_spectacular/plumbing.py
@@ -891,7 +891,7 @@ def modify_for_versioning(patterns, method, path, view, requested_version):
     elif issubclass(view.versioning_class, versioning.NamespaceVersioning):
         try:
             view.request.resolver_match = get_resolver(
-                urlconf=tuple(detype_pattern(p) for p in patterns)
+                urlconf=detype_patterns(patterns)
             ).resolve(path)
         except Resolver404:
             error(f"namespace versioning path resolution failed for {path}. Path will be ignored.")
@@ -976,6 +976,12 @@ def analyze_named_regex_pattern(path):
     return result
 
 
+@cache
+def detype_patterns(patterns):
+    """ Cache detyped patterns due to the expensive nature of rebuilding URLResolver """
+    return tuple(detype_pattern(p) for p in patterns)
+
+
 def detype_pattern(pattern):
     """
     return an equivalent pattern that accepts arbitrary values for path parameters.
```

> Given the nature of this, I'm wondering whether this might also partially address the issue raised in https://github.com/tfranzel/drf-spectacular/issues/597.

I still don't know how that happens, but it could only possibly be related if they were using `NamespaceVersioning`. This code path will only run if namespace versioning is encountered.

Would you mind sharing a performance plot of a big project? That would be awesome! It's hard to come by performance stats from other people.

```bash
apt-get install graphviz
pip install gprof2dot
gprof2dot -f pstats YOURPROFILE -z "generators:258:get_schema" --depth 15 | dot -Tsvg -o output.svg
```
You may need to fiddle with depth to get a usable plot. Also the line number `258` might be different for you.