@ngnpope 2022-03-20
> Your intuition is right about caching this, but I would be more at ease if it were slightly different. Caching it once (since it is not supposed to change), while not wasting memory on cached values that are never retrieved again. This should be identical in terms of speed assuming there are no reused subtrees. Could you try and benchmark this instead?

Sure. This makes sense. I tried it out and there was no discernible difference as expected so I incorporated it into the commit.
I had to pass `tuple(patterns)` because `patterns` is a list which is unhashable and doesn't work with `@cache`.

I have also added an extra commit to cache the call to `.resolve()`. Duplicate values of `path` can be provided where there are multiple methods (`GET`, `POST`, etc.) and I found a drop in the number of calls from 2495 to 1742. This doesn't seem to have any appreciable difference to performance either, but avoids duplicated effort.

> Would you mind sharing a performance plot of a big project? That would be awesome! It's hard to come by performance stats from other people.

I generated the plots as requested and found nothing identifying, so sure.

<details>
<summary><tt>000.svg</tt> -- original version</summary>

![000](https://user-images.githubusercontent.com/2855582/159184545-bbac0c2b-cae2-4308-8f5b-c7b8ca8d800b.svg)

</details>
<details>
<summary><tt>001.svg</tt> -- added <tt>@cache</tt> to <tt>detype_pattern()</tt></summary>

![001](https://user-images.githubusercontent.com/2855582/159184559-f54d6c63-df48-46e3-8d5d-171c5a352cd2.svg)

</details>
<details>
<summary><tt>002.svg</tt> -- changed to add <tt>detype_patterns()</tt> with <tt>@cache</tt></summary>

![002](https://user-images.githubusercontent.com/2855582/159184560-9d6623f5-f463-4d71-9180-1f3ab470c25f.svg)

</details>
<details>
<summary><tt>003.svg</tt> -- added <tt>resolve_path()</tt> with <tt>@cache</tt></summary>

![003](https://user-images.githubusercontent.com/2855582/159185712-ca900e99-6c42-4867-b85d-e894cde6b629.svg)

</details>