@tfranzel 2022-03-21
@ngnpope thank you so much. The plots look good. It's tricky getting good profiling indeed.

I chose to include only the first commit, since the second one had a negligible gain. The reason is we would cache resolver objects that are internal to DRF and I would like to keep the attack surface for dangling references as low as possible. This would again store a resolver objects for each and every path. Those are the things that potentially contribute to memory leaks. Probably not an issue here, but just to be on the safe side.