@tfranzel 2022-03-21
merging a partial PR leaves the PR open. the more you know :smile: Feel free to follow up on this if you like.