@ngnpope 2022-03-21
> I chose to include only the first commit, since the second one had a negligible gain. The reason is we would cache resolver objects that are internal to DRF and I would like to keep the attack surface for dangling references as low as possible. This would again store a resolver objects for each and every path. Those are the things that potentially contribute to memory leaks. Probably not an issue here, but just to be on the safe side.

Not a problem. Was more just to highlight an additional place with a significant number of extra calls.

I'm not sure if there is another way that we can avoid calling this with the same path for multiple methods without caching the resolver objects. It'd probably require too much rejigging of stuff around this line:

https://github.com/tfranzel/drf-spectacular/blob/6cdd515489417b91b64e0119c5d7319c5f51dd32/drf_spectacular/generators.py#L223

I checked and the `method` argument to `modify_for_versioning()` isn't used:

https://github.com/tfranzel/drf-spectacular/blob/6cdd515489417b91b64e0119c5d7319c5f51dd32/drf_spectacular/plumbing.py#L877

> merging a partial PR leaves the PR open. the more you know smile Feel free to follow up on this if you like.

Thanks! 😄